Kokkos:
Tune host thread binding algorithm to insure gangs using hyperthreads
will align with cores.  Share the mapping algorithm between the
pthreads and OpenMP devices.

+++ packages/kokkos/array/src/impl/Kokkos_Atomic_Fetch_Add.hpp
@@ -100,7 +100,7 @@ int atomic_fetch_add( volatile int * const dest , const int val )
 { return __sync_fetch_and_add(dest,val); }
 
 KOKKOSARRAY_INLINE_FUNCTION
-long long int atomic_fetch_add( volatile long int * const dest , const long int val )
+long int atomic_fetch_add( volatile long int * const dest , const long int val )
 { return __sync_fetch_and_add(dest,val); }
 
 #if defined( KOKKOS_ATOMICS_USE_GCC )
